refactor: split Pietta extension into focused modules
Break src/index.ts into dedicated core, commands, and tools modules
for git/worktree lifecycle, layout bootstrapping, memory operations,
context building, command parsing/registration, and tool registration.

Also document the new code layout in AGENTS.md.

Keeps behavior unchanged while making the extension easier to reason
about, test, and maintain.

Fixes https://github.com/bitbloxhub/pietta/issues/8
